Refactor NavigationSidebar and collapse components
Deleted ContentCollapse.tsx and refactored ManagementCollapse.tsx into a more generic NavigationSidebarCollapse component. This new component can handle different titles, icons and menu items, improving code reuse and maintainability. The changes also involved moving some functionality from the now-deleted ContentCollapse to the NavigationSidebar.tsx file.
