+++ frontend/src/components/NavigationSidebar/NavigationSidebar.tsx
@@ -12,18 +12,20 @@ import {
   StackDivider,
   Text,
 } from "@chakra-ui/react";
-import { ContentCollapse } from "./collapses/ContentCollapse.tsx";
 import { SidebarButton } from "./SidebarButton.tsx";
 import {
   FaArrowRightFromBracket,
   FaBorderAll,
+  FaBriefcase,
   FaEllipsisVertical,
+  FaFilm,
 } from "react-icons/fa6";
 import { logout } from "../../services/authService.ts";
 import { useNavigate } from "react-router-dom";
 import Logo from "../Logo/Logo.tsx";
-import { ManagementCollapse } from "./collapses/ManagementCollapse.tsx";
 import { GithubUserAuthType } from "../../model/githubModel.ts";
+import { useContentCreationDrawerStore } from "../../store/store.ts";
+import NavigationSidebarCollapse from "./NavigationSidebarCollapse.tsx";
 
 type NavigationSidebarProps = {
   user: GithubUserAuthType;
@@ -33,6 +35,34 @@ export default function NavigationSidebar({
   user,
 }: Readonly<NavigationSidebarProps>) {
   const navigate = useNavigate();
+  const onOpenContentCreationDrawer = useContentCreationDrawerStore(
+    (state) => state.onOpen
+  );
+
+  const contentCollapseItems = [
+    {
+      id: 1,
+      title: "All Content",
+      onClick: () => {
+        navigate("/content");
+      },
+    },
+    {
+      id: 2,
+      title: "Create content",
+      onClick: onOpenContentCreationDrawer,
+    },
+  ];
+
+  const managementCollapseItems = [
+    {
+      id: 1,
+      title: "Staff",
+      onClick: () => {
+        navigate("/staff");
+      },
+    },
+  ];
 
   return (
     <Flex as="section" minH="100vh">
@@ -55,8 +85,16 @@ export default function NavigationSidebar({
             >
               Dashboard
             </SidebarButton>
-            <ContentCollapse />
-            <ManagementCollapse />
+            <NavigationSidebarCollapse
+              title="Content"
+              icon={FaFilm}
+              menuItems={contentCollapseItems}
+            />
+            <NavigationSidebarCollapse
+              title="Management"
+              icon={FaBriefcase}
+              menuItems={managementCollapseItems}
+            />
           </Stack>
         </Stack>
         <Stack spacing="4" divider={<StackDivider />}>
