+++ frontend/src/components/NavigationSidebar/NavigationSidebarCollapse.tsx
@@ -9,22 +9,24 @@ import {
   useDisclosure,
 } from "@chakra-ui/react";
 import { FiChevronDown } from "react-icons/fi";
-import { FaBriefcase } from "react-icons/fa6";
-import { useNavigate } from "react-router-dom";
+import { IconType } from "react-icons";
 
-export const ManagementCollapse = () => {
-  const navigate = useNavigate();
+type CollapseProps = {
+  title: string;
+  icon: IconType;
+  menuItems: {
+    id: number;
+    title: string;
+    onClick: () => void;
+  }[];
+};
+
+export default function NavigationSidebarCollapse({
+  title,
+  icon,
+  menuItems,
+}: CollapseProps) {
   const { isOpen, onToggle } = useDisclosure();
-
-  const data = [
-    {
-      id: 1,
-      title: "Staff",
-      onClick: () => {
-        navigate("/staff");
-      },
-    },
-  ];
 
   return (
     <Box>
@@ -35,14 +37,14 @@ export const ManagementCollapse = () => {
         width="full"
       >
         <HStack spacing="3">
-          <Icon as={FaBriefcase} />
-          <Text as="span">Management</Text>
+          <Icon as={icon} />
+          <Text as="span">{title}</Text>
         </HStack>
         <PopoverIcon isOpen={isOpen} />
       </Button>
       <Collapse in={isOpen} animateOpacity>
         <Stack spacing="1" alignItems="stretch" ps="8" py="1">
-          {data.map((item) => (
+          {menuItems.map((item) => (
             <Button
               key={item.id}
               variant="tertiary"
@@ -56,7 +58,7 @@ export const ManagementCollapse = () => {
       </Collapse>
     </Box>
   );
-};
+}
 
 export const PopoverIcon = (props: { isOpen: boolean }) => {
   const iconStyles = {
